GET, PUT, DELETE /posts/{postId} 404 예외 처리 추가
PUT, DELETE, GET에 대하여 존재하지 않는 포스트에 대한 작업은 404를
리턴하도록 수정

+++ src/main/java/com/com/cnu/devlog_springboot/controller/PostController.java
@@ -24,7 +24,12 @@ public class PostController {
     // GET /posts/{postId}
     @GetMapping("{postId}")
     public ResponseEntity<Post> getPost(@PathVariable("postId")Integer postId) {
-        return ResponseEntity.ok(postService.getPost(postId));
+        Post post = postService.getPost(postId);
+        if (post != null) {
+            return ResponseEntity.ok(postService.getPost(postId));
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     // POST /posts
@@ -40,13 +45,23 @@ public class PostController {
             @PathVariable("postId")Integer postId,
             @RequestBody PostRequest postRequest
     ) {
-        return ResponseEntity.ok(postService.updatePost(postId, postRequest));
+        Post post = postService.getPost(postId);
+        if (post != null) {
+            return ResponseEntity.ok(postService.updatePost(postId, postRequest));
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
     // DELETE /posts/{postId}
     @DeleteMapping("{postId}")
     public ResponseEntity<Void> deletePost(@PathVariable("postId") Integer postId) {
-        postService.deletePost(postId);
-        return ResponseEntity.noContent().build();
+        Post post = postService.getPost(postId);
+        if (post != null) {
+            postService.deletePost(postId);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 }
